backend/requirements.txt
fastapi>=0.124.4
uvicorn>=0.38.0
sqlmodel>=0.0.27
psycopg2>=2.9.11
python-dotenv>=1.2.1
pydantic>=2.12.5
pydantic-core>=2.41.5
passlib>=1.7.4
python-multipart>=0.0.20
mcp>=1.0.0
jwt>=1.3.1
bcrypt>=4.0.1